refactor: move error component
